Windows XP rendering fix + Disable minimize button when autoclicking
Fixed all radio buttons rendering wrong on Windows XP. This was fixed by disabling visual styles on Windows XP.
The minimize button now gets hidden when the autoclicker gets activated. This is to avoid the situtation where a user has not set a keybind and activated the autoclicker. This could cause the user to struggle to kill the autoclicker.